Revert "Merge branch 'unifyai:master' into master"
This reverts commit 38856362563dc36a1c90dda06e98323e359b7a3a, reversing
changes made to 1c512c46a874d0f49e9eac736d7a241e07f77d5d.

+++ ivy/functional/backends/jax/random.py
@@ -57,12 +57,11 @@ def random_uniform(
     else:
         RNG_, rng_input = jax.random.split(_getRNG())
         _setRNG(RNG_)
+
     return to_device(
-        jax.random.uniform(
-            rng_input, shape, minval=low, maxval=high, dtype=jnp.float32
-        ),
+        jax.random.uniform(rng_input, shape, minval=low, maxval=high, dtype=dtype),
         device,
-    ).astype(dtype)
+    )
 
 
 def random_normal(
