+++ ivy/functional/backends/torch/random.py
@@ -34,9 +34,7 @@ def random_uniform(
         torch.manual_seed(seed)
     if torch.is_tensor(shape):
         shape = shape.tolist()
-    return (
-        torch.rand(shape, device=device, dtype=torch.float) * rand_range + low
-    ).type(dtype)
+    return torch.rand(shape, device=device, dtype=dtype) * rand_range + low
 
 
 def random_normal(
